fix: fix test error
1. fix compile error after updating cpuinfo.py
2. fix all test errors, get through 'make test'

+++ bcolz/carray_ext.pyx
@@ -1653,7 +1653,7 @@ cdef class carray:
             if isize == 0:
                 newshape = newshape[:pos] + (0,) + newshape[pos + 1:]
             else:
-                newshape = newshape[:pos] + (isize / osize,) + newshape[
+                newshape = newshape[:pos] + (int(isize / osize),) + newshape[
                                                                pos + 1:]
             newsize = np.prod(newshape)
